+++ components/row-col/index.ts
@@ -1,7 +1,7 @@
 import Col from "./col";
 import Row from "./row";
 
-export type { ColProps } from './col';
+export type { ColProps, ColResponsiveSize, ColSize } from './col';
 export type { RowProps } from './row';
 
 export { Col, Row };